PropertySourcesPlaceholderConfigurer.setPropertySources()
For maximum configurability, PSPC now exposes
a setPropertySources(PropertySources) method.  When set, the supplied
PropertySources shadow any local properties specified as well
as any Enviroment specified (via EnvironmentAware.setEnvironment()).

It is presumed that this method will not be used frequently, but rather
as an 'escape hatch' for users that need to perform highly specific
PropertySource customizations on a per-configurer basis.  Usually the
approach would be to customize the PropertySource order at the
Environment level once and only once at context construction time. If
this approach is insufficient, setPropertySources() gives complete
control.

This functionality will *not* be exposed through the <context:*/>
namespace, as it is not considered mainline functionality.  In
such cases, 'dropping down' to a bean definition should be considered
reasonable.